# Conflicts:
#	src/MEGASync/control/FileFolderAttributes.cpp

+++ src/MEGASync/gui/LowDiskSpaceDialog.h
@@ -14,13 +14,15 @@ class LowDiskSpaceDialog : public QDialog
     Q_OBJECT
 
 public:
-    explicit LowDiskSpaceDialog(qint64 neededSize, qint64 freeSize, qint64 driveSize,
-                                const DriveDisplayData& driveDisplayData,
-                                QWidget *parent = nullptr);
+    explicit LowDiskSpaceDialog(long neededSize,
+        long long freeSize,
+        long long driveSize,
+        const DriveDisplayData& driveDisplayData,
+        QWidget* parent = nullptr);
     ~LowDiskSpaceDialog();
 
 private:
-    static QString toString(qint64 bytes);
+    static QString toString(long long bytes);
     void setupShadowEffect();
 
     Ui::LowDiskSpaceDialog *ui;
