+++ src/MEGASync/gui/PlanWidget.cpp
@@ -102,7 +102,7 @@ void PlanWidget::updatePlanInfo()
         QString price (toPrice(mDetails.pricePerUserBilling / 100.,
                                mDetails.billingCurrencySymbol));
         smallPrice = tr("%1 per user").arg(price);
-        mUi->lMinimumUsersBusiness->setText(tr("minimum %n user", "", mDetails.minUsers));
+        mUi->lMinimumUsersBusiness->setText(tr("minimum %n user", "", static_cast<int>(mDetails.minUsers)));
     }
     else
     {
